Add new method in LoginController to deal with login failed.
The reason to use a login controller rather than just passing directly
to JSP page, is to have the power to add extra information, maybe log
the number of failed attempts that some user has with a username or
password, so it's possible do some intrusion detection.

Doing that way give us a lot of valuable things by having this outside
of just JSP pages and wired into a proper controller.